+++ CHANGELOG.md
@@ -3,11 +3,12 @@
 * Refactor IPC and worker process API.
     - **Breaking** Remove `zng::task::ipc`.
     - Add `zng::task::channel::ipc_channel` and related types.
-    - Add `zng::task::process` with the same worker process types. !!: TODO
+    - Add `zng::task::process` with the same worker process types.
     - IPC types are now also available in builds without `"ipc"` feature, internally patched to use normal channels.
     - **Breaking** Remove `zng::task::channel::{UnboundSender, UnboundReceiver}`.
     - **Breaking** Unified channel error types.
     - **Breaking** Remove conversions from underlying channel types.
+    - **Breaking** Remove public `duct` crate types.
 * Add blocking API for `zng::task::channel` sender and receiver.
 
 # 0.19.2